requirements.txt
greenlet==3.0.3
playwright==1.46.0
pyee==11.1.0
PyYAML==6.0.2
typing_extensions==4.12.2
